src/app/dashboard/face detection/emotion_recognition/requirements.txt
opencv-python>=4.5.0
tensorflow==2.15.0
numpy>=1.19.0
scikit-learn==1.3.2
matplotlib==3.8.2
pandas==2.1.3
pillow>=8.0.0
seaborn==0.13.0
keras==2.15.0
tqdm==4.66.1
mediapipe==0.10.8
dlib>=19.22.0
imutils>=0.5.4
scikit-image>=0.18.0 